backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.3
cryptography==47.0.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.13
passlib==1.7.4
pillow==12.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.12
PyYAML==6.0.3
qrcode==7.4.2
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
